refactor(storages): extract shared extension-storage helpers
The chrome and firefox adapters were line-for-line copies: same listener
registry, same fireStorageEvent, same createOnChanged, same assembly of
local/sync/managed. Only the access style differs — chrome is callback-based,
firefox is promise-based.

The previous commit made that worse by narrowing values in the chrome copy
alone, leaving firefox to break the same way the moment its typings tighten.

Shared parts move to utils/extension-storage: the per-area listener registry and
the value narrowing. Each adapter keeps its own registry instance, so a change in
one browser's storage still never reaches listeners registered on another's, and
each keeps only its genuine difference. Narrowing now applies to both.

Adapters shrink from 93 and 50 lines to 34 and 26.
