+++ src/storages/browser-storage.ts
@@ -1,42 +1,18 @@
-import { AsyncStorage, StorageChange, StorageChangeEvent, StorageChangeListener } from '../@types/storage'
+import { AsyncStorage } from '../@types/storage'
+import { Area, createListenerRegistry, toStorageChanges, toStoredItems } from '../utils/extension-storage'
 
-const listeners = {
-  local: new Set<StorageChangeListener>(),
-  sync: new Set<StorageChangeListener>(),
-  managed: new Set<StorageChangeListener>(),
-}
-
-type Area = keyof typeof listeners
-
-function fireStorageEvent(changes: { [key: string]: StorageChange }, area: Area) {
-  listeners[area].forEach(listener => {
-    listener(changes)
-  })
-}
+const registry = createListenerRegistry()
 
 browser.storage.onChanged.addListener((changes, area) => {
-  fireStorageEvent(changes, area as Area)
+  registry.fire(toStorageChanges(changes), area as Area)
 })
 
-function createOnChanged(area: Area): StorageChangeEvent {
-  return {
-    addListener(listener) {
-      listeners[area].add(listener)
-    },
-    removeListener(listener) {
-      listeners[area].delete(listener)
-    },
-    hasListener(listener) {
-      return listeners[area].has(listener)
-    },
-  }
-}
-
+// browser.storage is promise-based, unlike its callback-based chrome counterpart.
 const createStorage = (storage: browser.storage.StorageArea, area: Area): AsyncStorage => ({
-  get: storage.get,
+  get: async keys => toStoredItems(await storage.get(keys)),
   set: storage.set,
   remove: storage.remove,
-  onChanged: createOnChanged(area),
+  onChanged: registry.createOnChanged(area),
 })
 
 const local = createStorage(browser.storage.local, 'local')
